Ft/articles

+++ helpers/tools.go
@@ -6,10 +6,10 @@ import (
 )
 
 // ParseInt64 helper to avoid code repetition
-func ParseInt64(stringToParse string) (int64, error) {
+func ParseUInt64(stringToParse string) (uint64, error) {
 	intID, err := strconv.ParseInt(stringToParse, 10, 64)
 	if err != nil {
 		return 0, fmt.Errorf("could not parse string to int")
 	}
-	return intID, nil
+	return uint64(intID), nil
 }
